✨Middleware
In order to have an extensible architecture that will allow us to do
things like provide client capabilities that the upstream client does
not provide, we need a way to insert lspx into the conversation
between client and servers. An example of this is fixing the gap in
the conversation between `eglot` and `vscode-eslint-language-server`.

This applies the middleware pattern to all communication between
client and server so that a single middleware can intercept whichever
messages it needs to carry out its job. Because connections are
stateful and support both notification and request, there are four
components to each middleware:

client to N servers:

- notify
- request

1 of N servers to client:
- notify
- request

Because the middleware contains not only the request/notification
params, but also the lsp agents that are sending and receiving, it can
make decisions about filtering agents before forwarding on, or it can
ever respond directly without consulting the other end of the
connection.

As part of this, the lifecycle that captures capabilities has been
re-written as the very first middleware at the top of the stack, and
it actually makes it quite a bit cleaner. The lifecycle of the LSPX
server is a middleware that delegates to a "State" middleware.

This change limits itself to the minimum required to insert the
middleware into each piece of communication, but in the future we
should look to extracting out the special case "completion" handler in
the `defaultRequest` and implement it as middleware